[database:dump] Don't duplicate .gz extension (#3014)
* [database:dump] Don't duplicate .gz extension

* [database:dump] Keep the overwritten file

+++ src/Command/Database/DumpCommand.php
@@ -127,7 +127,9 @@ class DumpCommand extends Command
         if ($this->shellProcess->exec($command, $this->appRoot)) {
 						$resultFile = $file;
 						if ($gz) {
-							$resultFile = $file . ".gz";
+							if(substr($file, -3) != '.gz') {
+								$resultFile = $file . ".gz";
+							}
 							file_put_contents(
 									$resultFile,
 									gzencode(
@@ -135,7 +137,9 @@ class DumpCommand extends Command
 													$file)
 									)
 							);
-							unlink($file);
+							if($resultFile != $file) {
+								unlink($file);
+							}
 						}
 
             $io->success(
